requirements.txt
cycler==0.10.0
gcloud==0.17.0
googleapis-common-protos==1.53.0
httplib2==0.20.1
jws==0.1.3
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
oauth2client==3.0.0
pandas==1.3.3
Pillow==8.3.2
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pyparsing==2.4.7
Pyrebase==3.0.27
python-dateutil==2.8.2
python-dotenv==0.19.0
python-firebase==1.2
python-jwt==2.0.1
pytz==2021.3
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.7.2
scipy==1.7.1
seaborn==0.11.2
selenium==3.141.0
six==1.16.0
urllib3==1.26.7
